refactor: スキルパネルのスキル一覧表示部分をLiveViewComponentに変換
